Pulled the media api into this repository
twdoc uses an api from the /var/www/html/tw.rpi.edu/media directory with url http://tw.rpi.edu/media.

Also fixed the issue where the link to the generic document page wasn't using the latest uri. The reason for that is that that latest uri didn't exist. I
added all of the latest uris that were missing and changed the code to use that uri in the ?uri= query string.